refactor slider details into it's own component
